Added LinkedList class for managing Song objects.
Implemented a Node class within LinkedList to represent each element of the list, containing a Song object and a reference to the next Node.
Implemented a constructor to initialize the LinkedList with head, tail, and size variables.
Implemented an add method to append a Song to the end of the list.
Implemented a size method to retrieve the current size of the list.
Implemented a display method to print all songs in the list."